+++ ChangeLog.md
@@ -1,3 +1,11 @@
+Version 0.2:
+Add phases:
+     - Phys2VirtRegs - use virtual regs instead of physical
+     - DCE - dead code eliminations
+    Some auxiliary actions due to support this changes.
+
+
+
 Verson 0.1:
 Unfinished, unstable, not even MVP DSL-compiler
 Create some main compiler notions: